Make the SQL parser buildable outside the server, and cut its dependencies
Groundwork for building `src/Parsers` on its own (the WebAssembly experiment in
`utils/wasm-parser` that follows). Everything here stands on its own as a fix or a
build-time improvement; nothing is WebAssembly-specific except the platform guards.

Dependencies the parser did not need:

* `DumpASTNode.h` included `Poco/Util/Application.h` and never used it. That header
  reaches `Poco::ThreadPool` and `Poco::SignalHandler`, and `DumpASTNode.h` is pulled in
  by `InDepthNodeVisitor.h`, so over 60 translation units paid for it.
* `ExpressionListParsers.cpp` included `AggregateFunctions/AggregateFunctionFactory.h`
  and never used it either, dragging `IAggregateFunction` and all of `Columns` into one
  of the heaviest translation units in the tree.
* `ParserSystemQuery.cpp` included the whole Keeper client for three path-string
  helpers. Those move to a new `Common/ZooKeeper/ZooKeeperPathUtils.h`; `ZooKeeper.h`
  includes it, so every existing user is unaffected.

`ASTSQLSecurity::formatImpl` was defined identically in both `ASTSQLSecurity.cpp` and
`ASTCreateQuery.cpp`. Static-library linking never pulls both objects in, so the ODR
violation stayed latent; it becomes a duplicate-symbol error as soon as the objects are
linked directly. Removed the copy in `ASTCreateQuery.cpp`.

`CommonParsers.cpp` built the keyword table at startup: 645 inlined
`resize`/`std::string` assign/`unordered_set::emplace` sequences, which compiled to a
57 KB constructor - the single largest function in the parser - plus 645 string
allocations and a hash-set build in every process. It is a `constexpr
std::array<std::string_view, N>` now, indexed directly by `Keyword`; the two consistency
checks the constructor performed became `static_assert`s, so they now also run in release
builds. `getAllKeyWords` and `getKeyWordSet` only feed system tables and query
obfuscation, so they materialize on first use.

32-bit correctness, all of which were latent:

* `ASTs` asked Boost for a `uint32_t` stored size, which Boost rejects unless the type is
  strictly smaller than `size_t`. Pick `size_t` when there is nothing to shrink.
* `DEFAULT_NATIVE_BINARY_MAX_NUM_ROWS` was a `uz` literal that does not fit in a 32-bit
  `size_t`, and is compared against a `UInt64` read off the wire.
* `ProfileEvents::Count` was `size_t` while `Increment` was `Int64`, tripping the
  `static_assert` that the two match.
* `itoa` and `iota` guarded their `long`/`size_t` overloads on `OS_DARWIN`. The real
  condition is that `long` is a type of its own, which is also true on any 32-bit
  platform, so the guard is now `!defined(__LP64__)` as well.
* `CLOCK_MONOTONIC_COARSE` is a Linux extension; `base/time.h` listed the platforms that
  need a substitute one by one and now falls back generically.

`base/defines.h` and Poco's `Platform.h` learn about `__wasm__` so that headers naming
the platform stop erroring out.


Three visitors in `src/Databases` were relying on `DumpASTNode.h` to give them
`Poco/Util/AbstractConfiguration.h` transitively; they include it themselves now.

+++ src/Common/ProfileEvents.h
@@ -21,7 +21,8 @@ namespace ProfileEvents
 {
     /// Event identifier (index in array).
     using Event = StrongTypedef<size_t, struct EventTag>;
-    using Count = size_t;
+    /// Not `size_t`: counters are 64-bit even on 32-bit platforms, and must match `Increment`.
+    using Count = UInt64;
     using Increment = Int64;
 
     /// Counter cells are plain `Count`, accessed atomically via `std::atomic_ref`. Keeping the
